% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashPoissonModality.R
\name{ashPoissonWrapper}
\alias{ashPoissonWrapper}
\title{ash-Poisson wrapper for evaluating different bimodality scenarios}
\usage{
ashPoissonWrapper(counts, args = c(args.uni_natural, args.uni_zero,
  args.uni_natural_pm_zero, args.uni_natural_uni_zero))
}
\arguments{
\item{counts}{gene by sample count matrix}

\item{args}{pre-defined arguments for the four modality scenarios: args_uninonzero for unimodal distribution modeling after a mixture of half uniform distributions with mode at non-zero count, args_zero fo runimodal distribution modeling after a mixture of positive uniform distributionw ith mode at zero, and finally args_bimodal1 for a two-component situation where there's a zero-component modeling point mass at 0 and a non-zero component modeling a mixture of half uniform distribution with mode at 0.}
}
