% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsPreprocess.R
\name{filter.excludeAllZeros}
\alias{filter.excludeAllZeros}
\title{Filter all-zero samples and features.}
\usage{
filter.excludeAllZeros(counts)
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}
}
\value{
\code{counts} filtered count matrix.
}
