% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsPreprocess.R
\name{filter.Wrapper}
\alias{filter.Wrapper}
\title{Wrapper for all filtering steps}
\usage{
filter.Wrapper(counts, condition, is_nullgene = NULL)
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}
}
\value{
\code{counts} filtered count matrix.
}
