% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsPreprocess.R
\name{filterFeatures.fractionExpressed}
\alias{filterFeatures.fractionExpressed}
\title{Filter genes by fraction of samples detected as expressed}
\usage{
filterFeatures.fractionExpressed(counts, condition, thresholdDetection = 1,
  fractionExpressed = 0.01, control = list(plot = FALSE))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}

\item{thresholdDetection}{minimum count (or tpm or cpm, depends on the data) required to
qualified as "detected as expressed".}

\item{fractionExpressed}{fraction of samples qualified as detected as expressed.}

\item{control}{List with the following arguments
\code{plot} TRUE/FALSE to output diagnositic plot.}
}
\value{
List of the following objects
   \code{fraction_expressed_cond} fraction of samples qualified as "expressed" in each condition.
   \code{index_filter} TRUE/FALSE passing the filter.
}
