% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsPreprocess.R
\name{filterSamples.fractionExpressed}
\alias{filterSamples.fractionExpressed}
\title{Filter samples by fraction of genes detected as expressed}
\usage{
filterSamples.fractionExpressed(counts, thresholdDetection = 1,
  fractionExpressed = 0.25, control = list(plot = FALSE))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}

\item{thresholdDetection}{minimum count (or tpm or cpm, depends on the data) required to
qualified as "detected as expressed".}

\item{fractionExpressed}{fraction of genes qualified as detected as expressed.}

\item{control}{List with the following arguments
\code{plot} TRUE/FALSE to output diagnositic plot.}
}
\value{
List of the following objects
   \code{fraction_expressed} fraction of genes qualified as "expressed".
   \code{index_filter} TRUE/FALSE passing the filter.
}
