% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComp-blmer.r
\name{gene_variation}
\alias{gene_variation}
\title{Per gene variance component model}
\usage{
gene_variation(counts, annotation)
}
\arguments{
\item{annotation}{Meta-data matrix of each column of xx.}

\item{xx}{Matrix of expression measurements on log scale.}
}
\description{
Wrapper of the blmer function. Fits a bayesian nested model for one gene at a time.
}
\examples{
molecules_final <- get(load("molecules-final.txt"))
anno_filter <- get(load("annotation-filter.txt"))
blme_final <- gene_variation(counts = molecules_final,
                             annotation = anno_filter)
}
