% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsBenchmarks.R
\name{getAUC}
\alias{getAUC}
\title{Compute AUC}
\usage{
getAUC(response, predictor)
}
\arguments{
\item{response}{Binary indicator of true/false}

\item{predictor}{Numeric and continous variable.}
}
\description{
Compute AUC
}
