% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsBenchmarks.R
\name{getROC.average}
\alias{getROC.average}
\title{Plot ROC curve average}
\usage{
getROC.average(roc_output, length.out = 20)
}
\arguments{
\item{roc_output}{data.frame of roc output, including columns of true positive rate, 
false positive rate, and simulation dataset index.}

\item{length.out}{default 20. FPR is sorted and grouped into \code{length.out} internals.}
}
\value{
\code{output} data.frame of average true positive rates given false positive rates.
}
