% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsBenchmarks.R
\name{getTPR.pROC}
\alias{getTPR.pROC}
\title{Compute sensitivity or true positive rate given fixed FDR using pROC output}
\usage{
getTPR.pROC(response, predictor, fdr_cutoff = 0.05)
}
\arguments{
\item{fdr_cutoff}{default .05}

\item{pROC_output}{output from pROC package}
}
\value{
\code{tpr} true postive rate at the given FDR threshold.
}
