% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{getTPR}
\alias{getTPR}
\title{Compute sensitivity or true positive rate given fixed FDR}
\usage{
getTPR(response, predictor, fdr_cutoff = 0.05)
}
\arguments{
\item{response}{Binary indicator of true/false. True = Non-null gene and FALSE = Null gene.}

\item{predictor}{Significance value (e.g., lfsr or p-value).}

\item{fdr_cutoff}{False Discovery Rate threshold.}
}
\description{
Compute sensitivity or true positive rate given fixed FDR
}
