% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-singleCell.R
\name{methodWrapper.mast}
\alias{methodWrapper.mast}
\alias{methodWrapper.rots}
\title{MAST}
\usage{
methodWrapper.mast(count_matrix, condition, control = list(save_modelFit =
  FALSE, include_cdr = TRUE))

methodWrapper.rots(count_matrix, condition, control = list(save_modelFit =
  FALSE, seed = NULL, B = 100))
}
\arguments{
\item{count_matrix}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{include_cdr} MAST internal argument. TRUE if include cellular detection 
                   rate (fraction of genes detected per sample) as a covariate 
                   (scaled to mean 0 and standard deviation 1).}

\item{count_matrix}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{B} ROTS internal argument: Number of bootstraps. Default 100.
\code{seed} ROTS internal argument: An integer seed for the random number generator. 
            Default NULL.}
}
\value{
A list with the following objects
   \code{betahat} The estimate effect size of condition for all genes.
   \code{sebetahat} The standard errors of the effect sizes.
   \code{df} The degrees of freedom associated with the effect sizes.
   \code{pvalue} P-values of the effect sizes from the likelihood ratio test of 
                 the hurdle model, which combines the continous and the discrete 
                 component (the above betahat, sebetahat, and df are extracted from 
                 the continuous component.)
   \code{fit} MAST complete output of the model fit.

A list with the following objects
   \code{sig_order} Z-socres of expression difference, ordered from the most 
                    significant to the least significant.
   \code{fit} ROTS complete output of the model fit.
}
\section{Functions}{
\itemize{
\item \code{methodWrapper.mast}: MAST.

\item \code{methodWrapper.rots}: ROTS 1.2.0.
}}

\author{
Chiaowen Joyce Hsiao

Chiaowen Joyce Hsiao
}
