% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExpression.R
\name{make_normalmix}
\alias{make_normalmix}
\title{Generate beta (effects) from normal mixture prior}
\usage{
make_normalmix(ngene, pi, mu, sd, pi0)
}
\arguments{
\item{ngene}{Total number of genes (null + non-null in the data).}

\item{pi}{K-component probability vector. K = number of components in the mixture prior.}

\item{mu}{Length-K mean vector of the mixture components.}

\item{sd}{Length-K standard error vector of the mixture components.}

\item{pi0}{Proportion of null genes.Default = "random", which selects a random number between 0 and 1 from a uniform distribution.}
}
\value{
null Binary vector of length ngene. 1 = null, 0 = no null.

scalar value pi0 proportion of null
}
