% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExpression.R
\name{makeSimCount2groups}
\alias{makeSimCount2groups}
\title{Generate count matrix of all null genes}
\usage{
makeSimCount2groups(counts, Ngene = NULL, Nsamp,
  sample_method = c("per_gene", "all_genes"), control = list(pi0 = NULL))
}
\arguments{
\item{counts}{Gene expression count matrix from a dataset (usually consisting of one biological condition).}

\item{Ngene}{Number of genes in the simulated dataset.}

\item{Nsamp}{Number of samples per condition in the simulated data.}

\item{nullpi}{Proportion of null genes.}
}
