% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExpression.R
\name{makeSimCount2groups.sampleGenes}
\alias{makeSimCount2groups.sampleGenes}
\title{Generate count matrix for genes of different expression levels}
\usage{
makeSimCount2groups.sampleGenes(counts, Ngenes = NULL, Nsamples = 50,
  sampleGenes_method = c("high", "medium", "low"),
  sample_method = c("per_gene", "all_genes"))
}
\arguments{
\item{counts}{Gene expression count matrix from a dataset.}

\item{Ngenes}{Number of genes in the simulated dataset.}

\item{Nsample}{Number of samples in each condition.}

\item{pi0}{Proportion of null genes. Default to be 1.}
}
