% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-singleCell.R
\name{methodWrapper.bpsc}
\alias{methodWrapper.bpsc}
\title{BPSC}
\usage{
methodWrapper.bpsc(count_matrix, condition, control = list(save_modelFit =
  FALSE, estIntPar = FALSE, useParallel = TRUE))
}
\arguments{
\item{count_matrix}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{estIntPar} BPSC internal argument (TRUE/FALSE) for using only the expressed
                 genes to compute parameter estimates.
\code{useParalle} BPSC internal argument (TRUE/FALSE) for using code written for 
                 parallel computing.}
}
\value{
A list with the following objects
   \code{pvalue} significance value for the group effect.
   \code{fit} BPSC complete output of the model fit.
}
\description{
Implement BPSC 0.99.1 (not available via Bioconductor). This method
             requires that the input data is normalized (counts-per-million or
             fragments per-kilo-base per million).
}
\author{
Chiaowen Joyce Hsiao
}
