% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-bulk.R
\name{methodWrapper.DESeq2}
\alias{methodWrapper.DESeq2}
\title{DESeq2}
\usage{
methodWrapper.DESeq2(count_matrix, condition, control = list(save_modelFit =
  FALSE, independentFiltering = FALSE, cooksCutoff = FALSE))
}
\arguments{
\item{count_matrix}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{independentFiltering} DESeq2 internal argument. Default = FALSE to not apply gene filtering.
\code{cooksCutoff} DESeq2 internal argument. Default = FALSE to not exclude samples.}
}
\value{
A list with the following objects
  \code{betahat} The estimate effect size of condition for all genes.
  \code{sebetahat} The standard errors of the effect sizes.
  \code{pvalue} P-values of the effect sizes.
  \code{fit} DESeq2 complete output of the model fit. NULL if \code{save_modelFit = FALSE}.
}
\description{
Implement DESeq2.
}
\author{
Chiaowen Joyce Hsiao
}
