% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-bulk.R
\name{methodWrapper.edgeR}
\alias{methodWrapper.edgeR}
\title{edgeR}
\usage{
methodWrapper.edgeR(count_matrix, condition, control = list(save_modelFit =
  FALSE, norm.factors = rep(1:ncol(count_matrix))))
}
\arguments{
\item{count_matrix}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{norm.factors} Numeric vector of library size normalization factor.}
}
\value{
A list with the following objects
  \code{betahat} The estimate effect size of condition for all genes.
  \code{pvalue} P-values of the effect sizes.
  \code{fit} edgeR complete output of the model fit. NULL if \code{save_modelFit = FALSE}.
}
\description{
Implement edgeR.
}
\author{
Chiaowen Joyce Hsiao
}
