% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression.R
\name{methodWrapper.limmaVoom}
\alias{methodWrapper.limmaVoom}
\title{voom + limma}
\usage{
methodWrapper.limmaVoom(object, condition, W = NULL)
}
\arguments{
\item{object}{gene by sample expression matrix. Can be}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{log2counts}{G genes by N samples log2count matrix.}
}
