% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression.R
\name{methodWrapper.mast}
\alias{methodWrapper.mast}
\title{MAST}
\usage{
methodWrapper.mast(object, condition, cdr)
}
\arguments{
\item{object}{gene by sample expression matrix containing log2 adjusted counts.}

\item{condition}{vector specifying sample treatment condition.}

\item{cdr}{cellular detection rate, computed as \code{scale(colSums(counts > 0))} where counts is raw expression count matrix.}
}
\value{
p.value significance value from likelihood ratio test.
}
