% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsConfoundingCorrection.R
\name{methodWrapper.num_sv}
\alias{methodWrapper.num_sv}
\title{Estimate number of surrogate variables}
\usage{
methodWrapper.num_sv(object, condition)
}
\arguments{
\item{object}{gene by sample expression matrix. Can be}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{num_sv}{number of surrogate variables}
}
