% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsConfoundingCorrection.R
\name{methodWrapper.sva}
\alias{methodWrapper.sva}
\title{Surrogate Variable Analysis}
\usage{
methodWrapper.sva(object, condition, num_sv)
}
\arguments{
\item{object}{gene by sample expression matrix. Can be}

\item{condition}{Binary vector of length N indicating sample biological condition.}
}
