% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExpression.R
\name{non_null_sim}
\alias{non_null_sim}
\title{Simulate count matrix}
\usage{
non_null_sim(counts, args)
}
\arguments{
\item{counts}{G*N null count matrix, rows are genes and columns are samples}

\item{args}{List of arguments, including
- betaargs: parameters for the normal mixture distribution to generate signals
- pi0: null proportion. If pi0=="random" then pi0 will be randomly selected from U(0,1)}
}
