% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsNormalize.R
\name{normalize.census}
\alias{normalize.census}
\title{census}
\usage{
normalize.census(counts, control = list(save_modelFit = FALSE))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). Has to be TPM or FPKM.}

\item{control}{List with control arguments, including
\code{save_modelFit} TRUE to output the complete census output.}
}
\value{
List of the following objects
   \code{count_normed} numeric matrix of normalized count matrix, or also know as true
                       abundance by monocle.
   \code{libsize_factors} numeric matrix of scale factors for library sizes.
   \code{model_output} monocle complete output.
}
\description{
census is implemented in Monocle.
}
