% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsNormalize.R
\name{normalize.cpm}
\alias{normalize.cpm}
\title{Counts per million}
\usage{
normalize.cpm(counts, libsize_factors = rep(1, dim(counts)[2]),
  control = list(NULL))
}
\arguments{
\item{counts}{gene by sample expression count matrix (G by N).}

\item{libsize_factors}{numreic vector of the scale factors for library sizes.
Default to 1 - no adjustment.}

\item{control}{List with control arguments, including}
}
\value{
\code{cpm} numeric matrix of counts per million.
}
\description{
Compute CPM using the scale factors for library sizes.
}
