% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsNormalize.R
\name{normalize.lib}
\alias{normalize.lib}
\title{Library size normalization without adjustment}
\usage{
normalize.lib(counts, libsize_factors = rep(1, dim(counts)[2]),
  control = list(NULL))
}
\arguments{
\item{counts}{gene by sample expression count matrix (G by N).}

\item{libsize_factors}{numreic vector of the scale factors for library sizes.
Default to 1 - no adjustment.}

\item{control}{List with control arguments, including}
}
\value{
\code{libsize_factors} numeric vector of the scale factors for library size.
}
\description{
Scale factors equal 1 for all libraries
}
