% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsNormalize.R
\name{normalize.rle}
\alias{normalize.rle}
\title{RLE (relative log expression)}
\usage{
normalize.rle(counts, control = list(NULL))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}
}
\value{
\code{libsize_factors} numeric vector of the scale factors for library size.
}
\description{
DESeq was used to implement RLE. Note that we chose to use the implementation by
             the developer of the method. edgeR also implments a version of RLE, which gives
             different results: (calcNormFactors(method = "RLE"))
}
