% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsNormalize.R
\name{normalize.TMM}
\alias{normalize.TMM}
\title{TMM normalization}
\usage{
normalize.TMM(object)
}
\arguments{
\item{object}{Gene by sample expression count matrix.}
}
\value{
an edgeR \code{DGE} object containing the raw count expression matrix, the raw library size (\code{lib.size}) and the library size normalizing factor (\code{norm.factors}).
}
