% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExpression.R
\name{pois_thinning}
\alias{pois_thinning}
\title{Poisson thinning}
\usage{
pois_thinning(counts, condition, log2foldchanges)
}
\arguments{
\item{counts}{G genes by N samples null count matrix.}

\item{log2foldchanges}{Prior distribution of the betas (true effects).}
}
\value{
counts A simulated count matrix of non-null and null genes.
}
