% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.R
\name{query.evaluation}
\alias{query.evaluation}
\title{Evalute multiple normalization methods and multiple DE methods}
\usage{
query.evaluation(counts, condition, is_nullgene, methodsNormalize = c("TMM",
  "RLE", "census", "scran"), methodsMeanExpression = c("DESeq2", "limmaVoom"),
  report = c("fdr_cutoff_summary", "roc_plot"))
}
\examples{
ipsc_eset <- get(load(system.file("testdata", "HumanTungiPSC.rda", package = "ashbun")))
counts <- exprs(ipsc_eset)[sample(nrow(exprs(ipsc_eset)), ), ]

#---- generat simulated datasets
simdata_list <- simulationWrapper(counts, Nsim = 5, Nsample = 100, Ngene = 500)

#---- extract a single dataset as an example
#---- take pi0 = .9, the first simulated data
simdata <- simdata_list[[3]][[1]]

# ---- gather evaluation results
eval_ouptut <- query.evaluation(counts = simdata$counts,
                                condition = simdata$condition,
                                is_nullgene = simdata$is_nullgene,
                                methodsNormalize = c("TMM", "RLE", "census", "scran"),
                                methodsMeanExpression = c("DESeq2", "limmaVoom"),
                                report = "fdr_cutoff_summary")
           
}
\author{
Chiaowen Joyce Hsiao
}
