% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.simple.R
\name{query.evaluation.simple}
\alias{query.evaluation.simple}
\title{Evaluate multiple normalization methods and multiple DE methods}
\usage{
query.evaluation.simple(counts, condition, is_nullgene, methodsMeanExpression,
  report.control = list(fdr_control_threshold = 0.05), nsim = NULL)
}
\author{
Chiaowen Joyce Hsiao
}
