% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.R
\name{query.methodsMeanExpression}
\alias{query.methodsMeanExpression}
\title{Runs multiple DE methods}
\usage{
query.methodsMeanExpression(counts, counts_normed, condition, libsize_factors,
  methodsMeanExpression = c("DESeq2", "limmaVoom", "edgeR", "BPSC", "MAST",
  "ROTS"))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). Use filtered count data.}

\item{counts_normed}{Normalized expression count matrix 
(typicall CPM with normlized library size).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{libsize_factors}{Numeric vector of scale factors for library size factors.}

\item{methodsMeanExpression}{Chararacter vector of evaluted methods. To run all methods, use
c("DESeq2", "limmaVoom", "edgeR","BPSC", "MAST", "ROTS")}

\item{control}{\code{pseudocount} Default .5. For limmaVoom and MAST. If NULL, then do not use pseudocount.}
}
\value{
\code{pvalues} data.frame of significance values. Columns corresond to input methods.
}
\author{
Chiaowen Joyce Hsiao
}
