% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.simple.R
\name{query.methodsMeanExpression.simple}
\alias{query.methodsMeanExpression.simple}
\title{Runs multiple DE methods under their default setting}
\usage{
query.methodsMeanExpression.simple(counts, condition, is_nullgene,
  default = TRUE, methodsMeanExpression)
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). Use filtered count data.}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{methodsMeanExpression}{Chararacter vector of evaluted methods. To run all methods, use
c("DESeq2", "limmaVoom", "edgeR","BPSC", "MAST", "ROTS")}

\item{counts_normed}{Normalized expression count matrix
(typicall CPM with normlized library size).}

\item{libsize_factors}{Numeric vector of scale factors for library size factors.}

\item{control}{\code{pseudocount} Default .5. For limmaVoom and MAST. If NULL, then do not use pseudocount.}
}
\value{
\code{pvalues} data.frame of significance values. Columns corresond to input methods.
}
\examples{
ipsc_eset <- get(load(system.file("testdata", "HumanTungiPSC.rda", package = "ashbun")))
counts <- exprs(ipsc_eset)[sample(nrow(exprs(ipsc_eset)), 500), ]
condition <- pData(ipsc_eset)$replicate

----- Step 1: filtering
counts_filtered <- filter.excludeAllZeros(counts)
featuresToInclude <- filterFeatures.fractionExpressed(counts_filtered,
                                                     thresholdDetection = 1,
                                                     fractionExpressed = .01)$index_filter

samplesToInclude <-  filterSamples.fractionExpressed(counts_filtered,
                                                     thresholdDetection = 1,
                                                     fractionExpressed = .01)$index_filter

counts_filtered <- counts_filtered[featuresToInclude, samplesToInclude]

---- Step 2: compute library size factors
libsize_factors <- normalize.scran(counts = counts_filtered)$libsize_factors
counts_normed <- normalize.cpm(counts_filtered, libsize_factors)$cpm

---- Step 3: run DE methods
pvals_list <- query.methodsMeanExpression.simple(counts = counts_,
                condition = condition_filtered,
                default=TRUE,
                methodsMeanExpression = c("DESeq2", "edgeR", "limmaVoom","BPSC", "MAST"))
}
\author{
Chiaowen Joyce Hsiao
}
