% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.R
\name{query.methodsNormalization}
\alias{query.methodsNormalization}
\title{Run multiple normalization methods}
\usage{
query.methodsNormalization(counts, condition, methodsNormalize = c("TMM",
  "RLE", "census", "SCnorm", "scran"))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). Use filtered count data.}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{methodsNormalize}{Chararacter vector of evaluted methods. To run all methods, use
c("normalize.cpm", "normalize.tmm", "normalize.rle",
 "normalize.census", "normalize.scnorm", "normalize.scran")}
}
\value{
\code{libsize_factors} List of multiple size factors.
}
\author{
Chiaowen Joyce Hsiao
}
