% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.R
\name{query.pipeline}
\alias{query.pipeline}
\title{Evalute multiple normalization methods and multiple DE methods}
\usage{
query.pipeline(counts, condition, methodsNormalize, methodsMeanExpression)
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). 
Use raw count data before filtering.}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{methodsNormalize}{Chararacter vector of evaluted methods. To run all methods, use
c("normalize.cpm", "normalize.tmm", "normalize.rle",
 "normalize.census", "normalize.scnorm", "normalize.scran")}

\item{methodsMeanExpression}{Chararacter vector of evaluted methods. To run all methods, use
c("DESeq2", "limmaVoom", "edgeR","BPSC", "MAST", "ROTS")}

\item{null}{binary indicator of true/false. True = Non-null gene and FALSE = Null gene.}
}
\value{
\code{pvals_longformat} data.frame of pvals.
}
\author{
Chiaowen Joyce Hsiao
}
