% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExpression.R
\name{sampleingene}
\alias{sampleingene}
\title{Randomisation of sample at each gene}
\usage{
sampleingene(count_onegene, Nsamp)
}
\arguments{
\item{count_onegene}{Count vector of one gene.}

\item{Nsamp}{Total number of samples in the simulated data (both conditions included).}
}
