% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExpressionWrappers.R
\name{simulationWrapper}
\alias{simulationWrapper}
\title{Wrapper for simulating datasets}
\usage{
simulationWrapper(counts, Nsim, Nsample, Ngene = NULL)
}
\arguments{
\item{counts}{gene by sample count matrix}

\item{Nsim}{number of simulated datasets}

\item{Nsample}{number of samples per biological condition}

\item{Ngene}{number of genes. Defaults to include all genes in the input data.}
}
