% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateExpression.R
\name{simulationWrapper.sampleGenes}
\alias{simulationWrapper.sampleGenes}
\title{simulation wrapper with flexible gene sampling method}
\usage{
simulationWrapper.sampleGenes(counts, Nsim = 1, Nsamples = 50,
  Ngenes = NULL, pi0 = NULL, sample_method = c("all_genes", "per_gene"),
  sampleGenes_method = c("high", "medium", "low"),
  beta_args = args.big_normal(betapi = c(1), betamu = c(0), betasd = c(0.8)))
}
