% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-bulk.R
\name{voom.controlPseudocount}
\alias{voom.controlPseudocount}
\title{limma + voom}
\usage{
voom.controlPseudocount(counts, design, pseudocount = 0.5,
  pseudo_libsizes = 1, span = 0.5)
}
\arguments{
\item{counts}{gene by sample expression count matrix.}

\item{design}{design matrix, generated by R function model.matrix()}

\item{pseudocount}{default .5}

\item{pseudo_libsizes}{default 1.}
}
\value{
\code{w} Weights of dimension G by N.
}
\description{
Implement limma with voom and apply the empirical Bayes method
   in limma. This is a modified version of voom which allows to specify pseudocount and
   pseudo library size.
   Hence, voom: y <- t(log2(t(counts + pseudocount)/(lib.size + pseudo_libsizes) * 1e+06)).
}
\author{
Chiaowen Joyce Hsiao
}
