% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsBenchmarks.R
\name{getTPR}
\alias{getTPR}
\title{Compute sensitivity or true positive rate given fixed FDR}
\usage{
getTPR(true_positive_rate, false_positive_rate, fdr_cutoff = 0.05)
}
\arguments{
\item{true_positive_rate}{numeric vector of sensitivity.}

\item{false_positive_rate}{numeric vector of 1 - specificity.}

\item{fdr_cutoff}{default .05}
}
\value{
\code{tpr} true postive rate at the given FDR threshold.
}
