% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-singleCell.R
\name{methodWrapper.mast}
\alias{methodWrapper.mast}
\alias{methodWrapper.rots}
\title{MAST}
\usage{
methodWrapper.mast(counts, condition, default = FALSE, pseudocount = 0.5,
  control = list(save_modelFit = FALSE, include_cdr = TRUE))

methodWrapper.rots(counts, condition, control = list(save_modelFit = FALSE,
  seed = NULL, B = 100))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). MAST requires the default
input to be log2CPM. In this function, the required input is
normalized expression counts, such as CPM. log2CPM is computed internally.}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{pseudocount}{Default .5.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{include_cdr} MAST internal argument. TRUE if include cellular detection
                   rate (fraction of genes detected per sample) as a covariate
                   (scaled to mean 0 and standard deviation 1).}

\item{counts}{Gene by sample expression count matrix (G by N). ROTS requires this input
to be normalized expression matrix, such as FPKM or CPM.}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{List with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{B} ROTS internal argument: Number of bootstraps. Default 100.
\code{seed} ROTS internal argument: An integer seed for the random number generator.
            Default NULL.}
}
\value{
List with the following objects
   \code{betahat} The estimate effect size of condition for all genes.
   \code{sebetahat} The standard errors of the effect sizes.
   \code{df} The degrees of freedom associated with the effect sizes.
   \code{pvalue} P-values of the effect sizes from the likelihood ratio test of
                 the hurdle model, which combines the continous and the discrete
                 component (the above betahat, sebetahat, and df are extracted from
                 the continuous component.)
   \code{fit} MAST complete output of the model fit.

A list with the following objects
   \code{sig_order} Z-socres of expression difference, ordered from the most
                    significant to the least significant.
   \code{fit} ROTS complete output of the model fit.
}
\section{Functions}{
\itemize{
\item \code{methodWrapper.mast}: MAST.

\item \code{methodWrapper.rots}: ROTS 1.2.0.
}}

\author{
Chiaowen Joyce Hsiao

Chiaowen Joyce Hsiao
}
