% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-singleCell.R
\name{methodWrapper.bpsc}
\alias{methodWrapper.bpsc}
\title{BPSC}
\usage{
methodWrapper.bpsc(counts, condition, default = FALSE,
  control = list(save_modelFit = FALSE, estIntPar = FALSE, useParallel =
  TRUE))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). BPSC requires this input
to be normalized expression matrix, such as FPKM or CPM.}

\item{condition}{Binary vector of length N indicating sample biological condition.
#' @param libsize_factors Numeric vector of scale factors for library sizes.
                      Default NULL multiplies all library sizes by 1.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{estIntPar} BPSC internal argument (TRUE/FALSE) for using only the expressed
                 genes to compute parameter estimates.
\code{useParalle} BPSC internal argument (TRUE/FALSE) for using code written for
                 parallel computing.}
}
\value{
List with the following objects
   \code{pvalue} significance value for the group effect.
   \code{fit} BPSC complete output of the model fit.
}
\description{
Implement BPSC 0.99.1 (not available via Bioconductor). This method
             requires that the input data is normalized (counts-per-million or
             fragments per-kilo-base per million).
}
\author{
Chiaowen Joyce Hsiao
}
