% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-bulk.R
\name{methodWrapper.DESeq2}
\alias{methodWrapper.DESeq2}
\title{DESeq2}
\usage{
methodWrapper.DESeq2(counts, condition, libsize_factors = NULL,
  default = FALSE, control = list(save_modelFit = FALSE,
  independentFiltering = FALSE, cooksCutoff = FALSE))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{libsize_factors}{Numeric vector of scale factors for library sizes.
Default NULL computes libsize_factors using DESeq2.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{independentFiltering} DESeq2 internal argument. Default = FALSE to not apply gene filtering.
\code{cooksCutoff} DESeq2 internal argument. Default = FALSE to not exclude samples.}
}
\value{
List with the following objects
  \code{betahat} Estimate effect size of condition for all genes.
  \code{sebetahat} Standard errors of the effect sizes.
  \code{pvalue} P-values of the effect sizes.
  \code{fit} DESeq2 complete output of the model fit. NULL if \code{save_modelFit = FALSE}.
}
\description{
Implement DESeq2.
}
\author{
Chiaowen Joyce Hsiao
}
