% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-bulk.R
\name{methodWrapper.edgeR}
\alias{methodWrapper.edgeR}
\title{edgeR}
\usage{
methodWrapper.edgeR(counts, condition, libsize_factors = NULL,
  default = FALSE, control = list(save_modelFit = FALSE))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{libsize_factors}{Numeric vector of scale factors for library sizes.
Default NULL computes libsize_factors using edgeR.}

\item{control}{List with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.}
}
\value{
List with the following objects
  \code{betahat} Estimate effect size of condition for all genes.
  \code{pvalue} P-values of the effect sizes.
  \code{fit} edgeR complete output of the model fit. NULL if \code{save_modelFit = FALSE}.
}
\description{
Implement edgeR.
}
\author{
Chiaowen Joyce Hsiao
}
