% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-bulk.R
\name{methodWrapper.limmaVoom}
\alias{methodWrapper.limmaVoom}
\title{limma + voom}
\usage{
methodWrapper.limmaVoom(count_matrix, condition, control = list(save_modelFit
  = FALSE))
}
\arguments{
\item{count_matrix}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete limmaVoom fit results.}
}
\value{
A list with the following objects
   \code{betahat} The estimate effect size of condition for all genes.
   \code{sebetahat} The standard errors of the effect sizes.
   \code{df} The degrees of freedom associated with the effect sizes.
   \code{pvalue} P-values of the effect sizes.
   \code{fit} limmaVoom complete output of the model fit.
}
\description{
Implement limma with voom and apply the empirical Bayes method 
   in limma.
}
\author{
Chiaowen Joyce Hsiao
}
