% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-bulk.R
\name{methodWrapper.limmaVoom}
\alias{methodWrapper.limmaVoom}
\title{limma + voom}
\usage{
methodWrapper.limmaVoom(counts, condition, pseudocount = 0.5,
  default = FALSE, control = list(save_modelFit = FALSE))
}
\arguments{
\item{counts}{normalized gene by sample expression count matrix. (CPM)}

\item{condition}{binary vector of length N indicating sample biological condition.}

\item{pseudocount}{default .5.}

\item{control}{List with control arguments, including
\code{save_modelFit} TRUE to output the complete limmaVoom fit results.}

\item{libsize_factors}{Numeric vector of scale factors for library sizes.
Default NULL multiplies all library sizes by 1.}
}
\value{
List with the following objects
   \code{betahat} Estimate effect size of condition for all genes.
   \code{sebetahat} Standard errors of the effect sizes.
   \code{df} Degrees of freedom associated with the effect sizes.
   \code{pvalue} P-values of the effect sizes.
   \code{fit} limmaVoom complete output of the model fit.
}
\description{
Implement limma with voom and apply the empirical Bayes method
   in limma.
}
\author{
Chiaowen Joyce Hsiao
}
