% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-singleCell.R
\name{methodWrapper.scde}
\alias{methodWrapper.scde}
\title{SCDE}
\usage{
methodWrapper.scde(count_matrix, condition, control = list(save_modelFit =
  FALSE, n.cores = 4, bootstrap_times = 100))
}
\arguments{
\item{count_matrix}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{A list with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{filter_by_scde} SCDE internal argument. TRUE if apply SCDE built-in filtering criteria.
\code{bootstrap_times} SCDE internal argument. Default 100 bootstraps.}
}
\value{
A list with the following objects
   \code{sig_order} Z-socres of expression difference, ordered from the most 
                    significant to the least significant.
   \code{fit} SCDE complete output of the model fit.
}
\description{
Implement SCDE 1.99.1. This older version of SCDE is need for computers
              that are compatible with an older version of flexmix (flexmix_2.3-13; 
              for related discussions on this, see https://github.com/hms-dbmi/scde/issues/40).
              In addition, SCDE requires a minimum of 2000 genes to perform the analysis.
}
\author{
Chiaowen Joyce Hsiao
}
