% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsMeanExpression-singleCell.R
\name{methodWrapper.scde}
\alias{methodWrapper.scde}
\title{SCDE}
\usage{
methodWrapper.scde(counts, condition, control = list(save_modelFit = FALSE,
  n.randomizations = 100, n.cores = 4, min.size.entries = ncol(counts)))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{control}{List with control arguments, including
\code{save_modelFit} TRUE to output the complete DESeq2 fit results.
\code{n.randomization} SCDE internal argument: Nubmer of bootstraps.
\code{n.cores} SCDE internal argument: Nubmer of cores.
\code{min.size.entries} SCDE internal argument: Mininum number of genes to use when
                         determining expected expression magnitude during model fitting.}
}
\value{
List with the following objects
   \code{sig_order} Z-socres of expression difference, ordered from the most
                    significant to the least significant.
   \code{fit} SCDE complete output of the model fit.
}
\description{
Implement SCDE 1.99.2. This older version of SCDE is need for computers
              that are compatible with an older version of flexmix (flexmix_2.3-13;
              for related discussions on this, see https://github.com/hms-dbmi/scde/issues/40).
              Currently, this function uses the SCDE method for normalizatino and
              does not allow using other normalization factors.
}
\author{
Chiaowen Joyce Hsiao
}
