% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsConfoundingCorrection.R
\name{methodWrapper.sva}
\alias{methodWrapper.sva}
\title{Surrogate Variable Analysis}
\usage{
methodWrapper.sva(count_matrix, design_matrix, num_sv)
}
\arguments{
\item{count_matrix}{gene by sample expression matrix. Can be}

\item{design_matrix}{Binary vector of length N indicating sample biological condition.}

\item{num_sv}{number of surrogate variables}
}
