% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsNormalize.R
\name{normalize.tmm}
\alias{normalize.tmm}
\title{TMM}
\usage{
normalize.tmm(counts, control = list(NULL))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N).}
}
\value{
\code{libsize_factors} numeric vector of the scale factors for library size.
}
