% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.R
\name{query.evaluation}
\alias{query.evaluation}
\title{Evaluate multiple normalization methods and multiple DE methods}
\usage{
query.evaluation(counts, condition, is_nullgene, methodsNormalize = c("LIB",
  "TMM", "RLE", "census", "SCnorm", "scran"),
  methodsMeanExpression = c("DESeq2", "limmaVoom"),
  report.control = list(fdr_cutoff = 0.05), nsim = NULL)
}
\arguments{
\item{thresholdDetection}{minimum count per gene/sample. Default value = 1.}

\item{fractionExpressed}{fraction of samples expressed (above thresholdDetection).
Default value = .01.}
}
\examples{
ipsc_eset <- get(load(system.file("testdata", "HumanTungiPSC.rda", package = "ashbun")))
counts <- exprs(ipsc_eset)[sample(nrow(exprs(ipsc_eset)), ), ]

#---- generat simulated datasets
library(ashbun)
simdata_list <- simulationWrapper(counts, Nsim = 2,
                                 Ngenes = 100,
                                 Nsam = 20,
                                 sample_method = "all_genes",
                                 pi0 = .5,
                                 beta_args = args.big_normal(betapi = 1,
                                                             betamu = 0, betasd = .8))

#---- extract a single dataset as an example
#---- take pi0 = .9, the first simulated data
simdata <- simdata_list[[1]]

# ---- gather evaluation results
eval_ouptut <- query.evaluation(counts = simdata$counts,
                                condition = simdata$condition,
                                is_nullgene = simdata$is_nullgene,
                                methodsNormalize = c("TMM", "RLE"),
                                methodsMeanExpression = c("DESeq2", "limmaVoom"),
                                report = "fdr_cutoff_summary")

}
\author{
Chiaowen Joyce Hsiao
}
