% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.simple.R
\name{query.methodsMeanExpression.simple}
\alias{query.methodsMeanExpression.simple}
\title{Runs multiple DE methods under their default setting}
\usage{
query.methodsMeanExpression.simple(counts, condition, is_nullgene,
  default = TRUE, methodsMeanExpression)
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). Use filtered count data.}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{methodsMeanExpression}{Chararacter vector of evaluted methods. To run all methods, use
c("DESeq2", "limmaVoom", "edgeR","BPSC", "MAST", "ROTS")}

\item{counts_normed}{Normalized expression count matrix
(typicall CPM with normlized library size).}

\item{libsize_factors}{Numeric vector of scale factors for library size factors.}
}
\author{
Chiaowen Joyce Hsiao
}
