% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.R
\name{query.methodsNormalization}
\alias{query.methodsNormalization}
\title{Run multiple normalization methods}
\usage{
query.methodsNormalization(counts, condition, methodsNormalize = c("LIB",
  "TMM", "RLE", "census", "SCnorm", "scran"))
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). Use filtered count data.}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{methodsNormalize}{Chararacter vector of evaluted methods. To run all methods, use
c("normalize.cpm", "normalize.tmm", "normalize.rle",
 "normalize.census", "normalize.scnorm", "normalize.scran")}
}
\value{
\code{libsize_factors} List of multiple size factors.
}
\examples{
ipsc_eset <- get(load(system.file("testdata", "HumanTungiPSC.rda", package = "ashbun")))
counts <- exprs(ipsc_eset)[sample(nrow(exprs(ipsc_eset)), 500), ]
condition <- pData(ipsc_eset)$replicate

----- Step 1: filtering
counts_filtered <- filter.excludeAllZeros(counts)
featuresToInclude <- filterFeatures.fractionExpressed(counts_filtered,
                                                     thresholdDetection = 1,
                                                     fractionExpressed = .01)$index_filter

samplesToInclude <-  filterSamples.fractionExpressed(counts_filtered,
                                                     thresholdDetection = 1,
                                                     fractionExpressed = .01)$index_filter

counts_filtered <- counts_filtered[featuresToInclude, samplesToInclude]
condition_filtered <- condition[samplesToInclude]

---- Step 2: compute library size factors
sizefactors <- query.methodsNormalization(counts_filtered,
                                          condition = condition_filtered,
                                          methodsNormalize = c("TMM", "RLE",
                                                                "census","scran"))

}
\author{
Chiaowen Joyce Hsiao
}
