% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMethods.R
\name{query.pipeline}
\alias{query.pipeline}
\title{Evalute multiple normalization methods and multiple DE methods}
\usage{
query.pipeline(counts, condition, methodsNormalize, methodsMeanExpression)
}
\arguments{
\item{counts}{Gene by sample expression count matrix (G by N). 
Use raw count data before filtering.}

\item{condition}{Binary vector of length N indicating sample biological condition.}

\item{methodsNormalize}{Chararacter vector of evaluted methods. To run all methods, use
c("normalize.cpm", "normalize.tmm", "normalize.rle",
 "normalize.census", "normalize.scnorm", "normalize.scran")}

\item{methodsMeanExpression}{Chararacter vector of evaluted methods. To run all methods, use
c("DESeq2", "limmaVoom", "edgeR","BPSC", "MAST", "ROTS")}

\item{null}{binary indicator of true/false. True = Non-null gene and FALSE = Null gene.}
}
\value{
\code{pvals_longformat} data.frame of pvals.
}
\examples{
ipsc_eset <- get(load(system.file("testdata", "HumanTungiPSC.rda", package = "ashbun")))
counts <- exprs(ipsc_eset)[sample(nrow(exprs(ipsc_eset)), 500), ]
condition <- pData(ipsc_eset)$replicate

----- Step 1: filtering
counts_filtered <- filter.excludeAllZeros(counts)
featuresToInclude <- filterFeatures.fractionExpressed(counts_filtered, 
                                                     thresholdDetection = 1,
                                                     fractionExpressed = .01)$index_filter
                                                   
samplesToInclude <-  filterSamples.fractionExpressed(counts_filtered, 
                                                     thresholdDetection = 1,
                                                     fractionExpressed = .01)$index_filter
                                                     
counts_filtered <- counts_filtered[featuresToInclude, samplesToInclude]

---- Step 2: 
sizefactors <- query.methodsNormalization(counts_filtered, condition = condition,
                                          methodsNormalize = c("TMM", "RLE",
                                                                "census","scran"))

ipsc_eset <- get(load(system.file("testdata", "HumanTungiPSC.rda", package = "ashbun")))
counts <- exprs(ipsc_eset)[sample(nrow(exprs(ipsc_eset)), 500), ]
condition <- pData(ipsc_eset)$replicate

results <- query.pipeline(counts = counts,
                          condition = condition,
                          methodsNormalize = c("TMM", "RLE", "census","scran"),
                          methodsMeanExpression = c("DESeq2", "limmaVoom"))

}
\author{
Chiaowen Joyce Hsiao
}
